Refactors to make funcs recievers for *AppProject
